all refactored
