+++ include/shared_defs.h
@@ -190,6 +190,8 @@
 #define EXCLUDE_WTYPE_GET 15
 #define INCLUDE_WTYPE_SET 16
 #define INCLUDE_WTYPE_GET 17
+#define WIN_START_SET 18
+#define WIN_START_GET 19
 
 #define AUTO_INTERVAL_DELAY_GET 0
